+++ src/MD3Search.php
@@ -1,6 +1,5 @@
 <?php
 
-require_once 'MD3.php';
 require_once 'MD3TextField.php';
 
 /**