+++ src/MD3TextField.php
@@ -1,6 +1,5 @@
 <?php
 
-require_once 'MD3.php';
 
 /**
  * MD3TextField - Material Design 3 Text Field Components